+++ changelog.md
@@ -1,7 +1,10 @@
 
 ### Version 0.1.0-SNAPSHOT
- - [refactor] Moved pull request service into own module
+ - [feature] Added directive for colors on wallboard
+ - [refactor] added tests for ranking directive and adopted pullRequest directive tests for dashboard
+ - [bug] fix layout for wallboard
  - [feature] add layout for repo filter list
+ - [refactor] Moved pull request service into own module
  - [bug] user handling
  - [feature] Use mocha reporter for karma
  - [feature] PRT 39 sort PR list
@@ -22,4 +25,4 @@
  - [bug] Fixed Makefile
  - [initial] initial commit
 
--- Last change by: Michael Diodone <michael.diodone@devbliss.com> Thu Mar 19 10:25:58 CET 2015
+-- Last change by: Michael Diodone <michael.diodone@devbliss.com> Mon Mar 23 12:36:49 CET 2015